[FV] Fixed bug in the calculation of the FV integral for the discounted case (gamma < 1) in compute_fv_integral()
The bug was that the factor (1 - gamma^(dt+1)) was used inside the sum, instead of the correct (1 - gamma^dt). Note that the correctness of the latter factor is confirmed by the recovery of the FV discounted summation formula in two border cases: dt = 0 and dt = 1, which before was NOT recovered.

In addition to this, the value of the reward was taken out of the sum() expression to make computation faster.

Finally, in LeaFV, fixed a small bug in the _update_integral() method where I was NOT computing the integral update by including the reward at the state of interest x (assumed a terminal state). Note that this bug did NOT have any effect so far because all non-zero rewards considered in my tests were always equal to +1.