[FV] Made the particle reactivation process more flexible, in preparation for the grouped-based FV learning of value functions; removed deprecated functions in queues.py
This is a commit containing cherry-picked changes, so execution could be compromised (although in this case it is unlikely because the changes are quite isolated of the unchanged parts).

Goal: Be able to choose the reactivation particle from a *subset* of the particles in the FV system, which in the grouped-based FV estimation will be the particles that belong to the same group identified by their common start state.
This is achieved by a new parameter called all_idx_particles which lists the particle IDs among which the reactivation particle can be selected.

In addition to this change, I refactored the absorption_number parameter as follows:
 - renamed it to --> reactivation_number because the former name did NOT convey what the variable actually contained, as it suggested that it was a counter of the number of absorptions experienced by the FV system, which is NOT the case. Instead the new name conveys that the particle ID to which the absorbed particle needs to be reactivated is already defined by this parameter.
 - changed its behaviour: now when this parameter reactivation_number is not None, the ROBINS reactivation method is automatically used.
This change of behaviour implied that I had to change the calls to reactivate_particle() in a few places in the code base, namely in discrete.py and queues.py.

Finally, I deleted the deprecated_fuctions_moved_to_estimators_and_simulators_fv() internal function in queues.py --> run_simulation_fv() which contained functions that were no longer called as they were moved to estimators.py or simulators.py, as indicated in the function's name.