[Simulator] Adapted the time step counter to start at 1 for the first step
Goal: Follow the usual convention in Markov chains where the time step is the time at which the Markov chain is at the state AFTER the transition. See e.g. Pierre Bremaud's book, Chapter 13, in particular the graph of the jump rate process.
This change also helps make the interpretation of the trajectories (e.g. when debugging) more intuitive, as the number of steps used in the cycle starts at 1 for the first step.

The change affects mostly the _run_single() and _run_single_continuing_task() methods of the discrete.Simulator class, as this convention of identifying the first step by t=1 had already been implemented in all three FV methods.
The change affects:
- the computation of the cycle time (when `set_cycle` is not None).
- the condition to end the simulation (based on parameter `max_time_steps`).
- the calculation of the number of time steps used during the simulation (which changed from `t+1` to `t` in the `return` line).

All 51 unit tests pass.