[ActorCritic] Generalized the value to set as initial policy so that learning becomes interesting also in 1D gridworlds (not only 2D)
In addition: Improved the calculation of the stationary distribution of an MDP so that it doesn't break when the Markov chain is periodic (e.g. 1D gridworld with an even number of states): now the stationary distribution is computed using the eigen-decomposition of the as opposed to from a large enough power of the transition probability matrix P (which gives different results in the periodic case depending on the power value, i.e. as it varies between `k*d` and `k*d + (d-1)`).

This is a partial commit so there may be a few things that do not work correctly.

DETAILS:
- policies/parameterized.py:
  - PolNN class: Generalized the policy initialization in neural-network-parameterized policies to any initial policy (not only a random policy as was the case before).