[FV] Implemented the iterative learning of the stationary probabilities and of the average reward
Goal: Learn in an online fashion the expected reward that is used as correction of the value functions and of the advantage function.

This is final commit of a series of partial commits. The two FV tests on the differential value functions in tests_estimators_discretetime.py fail because the current expected state and action values are based on a *non-iterative* estimation of the average reward (which affects the estimation of the differential value functions).
For now I do NOT update the expected results of these tests because I am not sure yet whether I will keep the previous tests, i.e. a test where the average reward is not estimated iteratively.

DETAILS:
- simulators/discrete.py:
  - Simulator class:
    - _run_simulation_fv(): Implemented the iterative estimation of the stationary probabilities and of the average reward. These estimates are updated whenever a new absorption event occurs in a particle that hasn't been absorbed before (i.e. an absorption event that is used for the estimation of the survival probability).