[FV] Added a check of whether the FV system is stuck at the same configuration
Goal: Be warned of the situation I've seen a few times where the state of the FV system doesn't change for several time steps. This is more likely to happen when the policy is close to deterministic.

The check is only added in _run_simulation_fv() in discrete.Simulator, NOT in the other FV simulator methods in the class (e.g. _run_simulation_fv_fraiman(), etc.).