[FVAC] Implemented replications with results from different value function learning methods (TD, FV) stored in dictionaries
This is a partial commit so there may be a few things that do not work correctly.
A CAREFUL SELECTION of hunks and lines to commit was carried out using SourceTree (which allows a more flexible selection of what to commit than PyCharm).

Summary of changes in the FVAC section of tests.py:
1) In terms of the simulation process:
- 2D arrays of size `nrep x n_learning_steps` were created to store the results of interest (e.g. loss_all, R_all, etc.) across replications for the currently executed policy learning method.
- Dictionaries were created to store those results of interest obtained for the different value function learning methods (e.g. FV, TD).
Ex: dict_loss['values_fv'] stores the loss values across replications and for each policy learning step when FV is used to learn the value functions.

2) In terms of the evaluation plots:
- The piece of code in tests.py that creates the plots used to evaluate and compare the different learning methods was updated to accomodate the process to the new data structure based on dictionaries.
- The *last* replication is used in the plots that work on just one replication.
- The ALTOGETHER plots section has been extended to include the plot of summary statistics on the replications run, namely the mean, median, standard error of mean, min, max computed across replications for each policy learning step.