[ActorCritic] Disentangled the value functions learner stored in LeaActorCriticNN from the one passed to the constructor
Goal: Avoid modifying the value functions learner of the critic when learning the policy using a critic.
Note that the value functions learner stored in the LeaActorCriticNN object is modified when the object is reset by the reset() method. And this resets valuable information that is potentially stored in the critic, such as the state visit count (which we may want to plot at some point).

The disentanglement was done by storing in the LeaActorCriticNN object a *copy* of the learner passed in parameter `learner_value_functions`.
The reset of the value functions learner done by reset() now also resets the average reward stored in the learner when parameter reset_value_functions=True.