[Environments] Rewards passed to EnvGridworld2D are now honoured, instead of being set only if the reward state is a terminal state
Goal: Consider more general reward landscapes, as e.g. negative rewards of moving to an obstacle, if that would make sense at some point.

This change implied fixing the check in the estimate_expected_reward() function in estimators/fv.py that verifies whether all non-zero reward states are present in the stationary probabilities dictionary that is used to compute the expected reward under stationarity. In fact, it happened already that some states whose stationary probability is NOT estimated by the FV simulation appear in the rewards landscape of a 2D gridworld with reward equal to 0.