[REFACTOR] Refactored the 1D gridworld environments defined in gridworlds.py and a few tests using them in test_estimators_discretetime.py
This is a partial commit so there may be a few things that do not work correctly.
Out of the 56 unit tests, all should pass except two FV tests on the 2D gridworld with obstacles in test_estimators_discretetime.py:

Goal: Simplify class landscape as EnvGridworld1D had become mostly redundant of the most developed and more flexible EnvGridworld1D_OneTerminalState class, which has evolved to accept:
- any set of terminal states
- any rewards landscape
which is not the case with the EnvGridworld1D class.

Therefore, I did the following changes:
- Renamed EnvGridworld1D --> EnvGridworld1D_Classic
- Renamed EnvGridworld1D_OneTerminalState --> EnvGridworld1D
- Made EnvGridworld1D_Classic inherit from EnvGridworld1D

In addition, in the newly named EnvGridworld1D class:
- Changed the default reward_default=1.0 to reward_default=0.0, as this is the case that is used in tests.
- Changed the default rewards landscape from being 0.0 at the terminal state to being 1.0 at the terminal state.
- Removed the specification of state s=0 as a transient state from the definition of the environment, as this is responsibility of the policy of the agent interacting with the environment.

Tests in the following files were updated to the new gridworld definitions:
- estimators/fv.py
- simu_lambdas.py
- test_estimators_discretetime.py
- test_TD_lambda.py
- test_FV.py
- test_simulators.py
- utils/computing.py

Finally, in test_estimators_discretetime.py, I renamed the test classes in order to make clear where learning on 1D gridworld environments vs. learning on 2D gridworld environemnts is tested.