[FV-learning] Experiments on the 1D gridworld using different average reward values (zero, true, etc.); new percentile() function
The experiments leverage a recently introduced parameter in the simulator methods of discrete.Simulator: use_average_reward_stored_in_learner
The CONCLUSION of these experiments is that there is indeed an effect of the average reward on the estimation accuracy of the state value function V(s).
However, the effect seems to be the oppositve between the FV and and the TD learner... (STRANGE)

In the TD case, an under/over-estimation of the average reward implies an under/over-estimation of the state value function V(s) for the states closer to the reward state.
In the FV case, the behaviour is EXACTLY the opposite...

The tested average reward values on the 5-state 1D gridworld (0-4) (with absorption set A = {1}, policy = [0.7, 0.3], N = 100, T = 2000, infinite time allowed for the FV system until all particles are absorbed at least once) are:
- 0.005 < true average reward = 0.015
- 0.035 > true average reward = 0.015

Results are reported in the RL-MeetingMinutes.docx document for entry date Mon, 22-Apr-2024.

Note that the current state of the tests.py code (where experiments were carried out) is the original one, i.e. where the average reward is estimated by the learner and not imposed externally. However, the parts that are needed to perform the above experiments are commented out or disabled.

Other changes performed:
- The seed used for the TD learner is the same as the seed used for the FV learner, to reduce the number of variability sources.
- The computation of the percentiles of the estimated V(s) on several replications is now done using the newly defined function in utils.computing: percentile()
- utils/computing.py: new function percentile() that can be used to compute the percentile. See example of use in tests.py where percentiles are computed using the groupby().agg() function of a pandas DataFrame.