[FVAC] New 2D gridworld characteristics can be set in tests.py: exit state, connected active set, wind
This is a partial commit so there may be a few things that do not work correctly.

The following environment characteristics can now be set:
a) exit_state_at_bottom --> used to create a small 2D labyrinth where we can easily (i.e. with a short simulation) see the advantage of FVAC over TDAC when the reward is rarely observed.
b) connected_active_set --> used to guarantee that the active set in FV is a connected set when the start state is part of the absorption set. The guarantee is given by removing the leftmost obstacle in the usual 2D labyrinth with corridor. This is done to prevent the problem already encountered where all FV particles remain distributed on the same part of the active set that is *disconnected* from the other part. In the maximal set A case of the usual 2D labyrinth with corridor, this implies that the FV system gets stuck forever at the state above the start state.
c) wind_dict --> used to define an environment with wind. Currently this is a stochastic environment, where the probability of being deviated from the deterministic destination cell after taking an action is given by the wind intensity, a value between 0 and 1.

NOTE 1: The introduction of WIND implies that the MAX average reward (i.e. the average reward under an optimal policy) is not as easily computed as in a labyrinth without wind (because the next state under the optimal deterministic policy of going through the corridor is no longer deterministic). Therefore the MAX average reward is now computed as the expected reward (a.k.a. true average reward) under stationarity and under the optimal deterministic policy of going through the corridor. The expected reward is computed using the transition matrix P of the labyrinth under a given policy which is used to compute the stationarity probability, which is in turn used to compute the expected reward under stationarity. This is done by the new function defined in tests.py called compute_true_state_value_function() which, as its name states, also computes the true state value function V(s).
From the MAX average reward computed in this way (which is the average reward for the CONTINUING learning task), an *approximate* value of the EPISODIC average reward is computed using the same formula that takes the MAX (continuing) average reward computed in the case without wind to the MAX episodic average reward, namely by multiplying the former by the factor L / (L-1) where L is the number of states in the corridor path. This is an approximation because the presence of wind makes this formula not valid any more, but it is a good enough approximation in terms of generating informed plots of the episodic average reward observed during the policy learning process.

NOTE 2: The creation of the 2D gridworld with WIND in test_optimizers_discretetime.py had been (INCORRECTLY) committed in commit 9483064 because the `wind_dict` parameter had not been defined in the Test_EstPolicy_EnvGridworldsWithObstacles class at that time. This means that the execution of the FVAC process on that commit will fail with an error that the `wind_dict` variable does not exist.