[Simulator] Ensured reproducibility by setting the seed of the policy's environment
Goal: Guarantee repeatability of results when the environment stored in the policy, although representing the same environment being simulated, has a *different memory address* than the environment stored in the discrete.Simulator object. If this is the case, the seeding done by the _run*() methods on the latter environment does NOT affect the choice of the next action done by the policy when the policy is stochastic. In fact, for stochastic policies, a random number is drawn using the random number generator stream associated to the *environment stored in the policy*, NOT the environment stored in the discrete.Simulator object. If they refer to the same environment, we are in business, but if not, the random numbers generated by the policy's environment's random stream would not be reproducible... unless we set its seed.

This is a partial commit so there may be a few things that do not work correctly.

DETAILS:
- simulators/discrete.py: Simulator class
  - In each of the 6 _run*() methods, the seed of the agent's policy's environment is set to the given seed. This guarantees that the actions chosen by the policy are reproducible.
  Note that the policy's environment is normally the same as the environment stored in the discrete.Simulator object, but it may be different if the policy has been previously copied without keeping the same reference to the environment stored in the original policy.
  - _run_single(), _run_single_continuing_task(): we now ALSO set the numpy seed by calling `np.random.seed(seed)`. This is relevant when parameter epsilon_random_action=True, as this requires generating random numbers with numpy in order to, firstly decide whether a random action needs to be generated, and secondly to generate a random number if that should be the case.
  IMPORTANT: This implies that, before this change, the results obtained by the TDAC learner when epsilon_random_action > 0 may not have been reproducible.

- agents/policies/parameterized.py: PolNN class:
  - reset(), init_policy(): new parameter `seed` that can be used to set the torch seed before generating the randomly perturbed weights with `nn.init.normal_()`. The seed is only set when it is not None (as `None` means that we want to keep using the stream that has potentially been started by a seeding already performed).

- agents/learners/value_functions.py: improved reproducibility of the RANDOM value function reset of a LinearValueFunctionApprox object by setting the seed ONLY when it is not `None` (because, if we set the seed when it is None, the generated random values will NOT be reproducible (as the seed becomes `None`!), and we want that `seed=None` indicate we want to keep using the random number generation stream that has potentially been started by a previously performed seeding).