[FVAC] Updated default execution parameters and environment configuration in tests.py
The following changes were done:
- the left and top margins of the absorption set A in the 2D labyrinth with corridor has been further reduced to occupy less space of the set of uninteresting states in order to make the problem a little more difficult for FV.
- parameter `max_time_steps_fv_per_particle` scales with the size of the active set and is set to 5*|B|, where B is the active set. This intends to incorporate a measure of the difficulty level of the value functions learning problem.
- parameter `max_time_steps_per_policy_learning_episode` has been increased 5-fold on 2D problems, from |S| to 5*|S|, where S is the state space. This parameter is used when performing the environment excursion that is responsible for computing the loss that is minimized by the Actor-Critic policy gradient method.

Note: Regarding the last parameter mentioned, `max_time_steps_per_policy_learning_episode`, I think introduces a new variable that makes the comparison between FVAC and TDAC more difficult because the larger its value, the easier it is to more thoroughly explore the environment, thus making learning the policy easier. This is true for any method that had been used to estimate the Critic, either FV or TD.
I think a better approach for comparing FV and TD is to use an offline learning of the policy, by computing the loss using the state visit probabilities estimated from the FV or TD excursions used to learn the value functions (Critic). When we do so, this `max_time_steps_per_policy_learning_episode` parameter will no longer be necessary and we thus simplify the method comparison process.