[FV] Fixed estimation of absorption set A due to misalignment of states and rewards history
SUMMARY: The estimation of the absorption set A had a bug coming from the misalignment of states and rewards stored in the learner trajectory.

CAUSE: Because rewards are stored differently by the GenericLearner class (used for a generic value function learner, where an initial dummy reward is not stored in self.rewards) and by the Learner class (used for value function learners in EPISODIC learning tasks, where an initial dummy reward is stored in self.rewards), the resulting trajectory of states, actions, rewards was misaligned, because the run_exploration() method used to collect the exploration trajectory uses the trajectory attributes (self.states, self.actions, self.rewards) stored in the GenericLearner class, and NOT the trajectory attributes (self._sttes, self._actions, self._rewards) stored in the Learner class.
This made reward R[k] (i.e. self.rewards[k]) be associated to state S[k] (i.e. self.states[k]), although in the actual Markov chain trajectory reward R[k] is associated with state S[k+1], i.e. R[k] is received when the Markov chain visits states S[k+1]. This misalignment made states with NON-ZERO reward (e.g. terminal states in gridworlds) be sometimes assigned to the estimated absorption set, and this is WRONG as no receiving-reward state should be present in the absorption set (at least for now).

SOLUTION: Updated the discrete.Simulator.run_exploration() method in order to correctly align the states and rewards observed during the Markov chain excursion.

This is a partial commit so there may be a few things that do not work correctly (although chances are rather low).

DETAILS:
- discrete/__init__.py:
  - _reset_at_start_of_episode(): replaced the dummy value stored at the self._rewards attribute when resetting the episode information from 0 to the reward received by the agent when visiting the start state of the episode. In our current application of estimating the absorption set A for Fleming-Viot, this avoids having a state with NON-ZERO reward incorrectly ending up in the distribution of visit frequencies to states with ZERO reward.

- discrete.py: Simulator class
  - run_exploration(): A dummy reward was added to the trajectory stored in the learner at the beginning of the exploration process and the final states and actions were added at the end of the process. This guarantees that retrieving reward learner.rewards[k] gives us the reward observed when the Markov chain visits state learner.states[k].