[Models] Backpropagation neural network can now have NO hidden layer
Goal: Use a backpropagation neural network to implement a policy driven by action preferences, such as those dictated by the Advantage function A(s,a). This would allow us to choose the action directly following the estimated Advantage function as opposed to by a theta parameter estimated with an Actor-Critic policy gradient methodology. The problem with this approach might be that the policy might be "too" stochastic, i.e. it may not converge to a deterministic policy as is the case with the parameterized policy learned with the Actor-Critic policy gradient method.

Note: A few files were changed just because the backpropagation class was renamed from nn_backprop to NNBackprop.

This is a partial commit so there may be a few things that do not work correctly (although chances are rather low).

DETAILS:
- estimators/nn_models.py:
  - Renamed nn_backprop class --> NNBackprop to comply by the PEP naming convention for classes.
  - NNBackprop: new method set_params() which sets all parameters in the neural network to the given values. This will be used later on to set the weights of a neural network model with no hidden layer to the advantage function A(s,a) in order to define a policy computed from the preferences for each action defined by the advantage function at each state. For now it is used as part of the new unit tests in test_estimators_models.py.

- policies/parameterized.py:
  - PolNN:
    - init_policy(): the behaviour of this method is now different depending on whether the neural network has or does not have a hidden layer. When no hidden layer is present, the given values of the desired initial policy are obtained EXACTLY (as opposed to approximately, when the neural network has a hidden layer), and this is achieved by setting the weights directly connecting the input layer to the output layer and setting the biases of the output neurons to zero.

- test/test_estimators_models.py: new file that tests neural network backpropagation architectures before allowing that the neural network has no hidden layer, in order to perform the change in the code more confidently. Now the file contains only one class (Test_Class_nn_backprop) and one test (test_model_architecture), which actually has three subtests.