[FV-patch] Fixed an error with the FV tests run in test_estimators_discretetime.py following commit 31000452 on 19-May-2024
The error was generated by not having renamed parameter `max_time_steps_fv` to `max_time_steps` following the parameter rename carried out on the _run_fv*() methods of discrete.Simulator.
The tests now define the value of the two parameters now received by the _run_fv*() methods: `max_time_steps` and `max_time_steps_for_absorbed_particles_check` which are set to the same value. This might be relaxed and we might only need to set one of those parameters but at this point I don't want to risk or spend time figuring that out.

At this point the test on the MountainCar in test_estimators_discretetime.py should fail (will be fixed in the next commit).