[FV] Loss network: new parameter J to *directly* specify the multidimensional sizes of the absorption set
This is a partial commit, so correct execution is not guaranteed.
However all 57 unit tests should pass except the usual 2 FV tests (see previous commit).

Goal: Avoid having to always go through the definition of the J/K factor (J_factor) to decide the value of J for each dimension of the problem. In fact, it is not always the case that the value of the absorption set in each dimention, J(i), should be a function of the blocking size for the respective dimension, K(i).

Summary: A new parameter `J` was added to both the execution of run_FV_LossNetwork.py and of the analyze_convergence() function defined therein that allows to DIRECTLY define the size of the absorption set in the different dimensions of the parameterized problem (e.g. `J=[2,3]`) as opposed to defining it indirectly through the values of the J/K fractions. This allows for a more intuitive definition of the J values, especially in the FV estimation process, where it is not always true that the J values must depend on the blocking size values, K. Recall that the accurate estimation of the expected reabsorption time, E(T_A), does not depend on K.
Note: Parameter `J` has priority over `J_factor`.

DETAILS:
- run_FV_LossNetwork.py:
  - New execution parameter J which hs been added to the existing one J_factor so that we can directly define the J values for each job clas as an integer value, instead of an indirect calculation via a J/K factor. If this parameter J is given, it has precedence over the J_factor-based calculation of J.

  - Improved the plot of the estimated probability and of the estimated expected cost *by blocking state* by including the states as labels. This is done at the end of the script.

- test/test_estimators_continuoustime.py: new assertions that check the execution parameter values.