[Bug] Fixed the calculation of the true stationary distribution for the single server system which could previously sum up to more than 1
This is a partial commit, so correct execution is not guaranteed.
Nevertheless, all unit tests should pass (as tested as if the commit had been a full commit), except for the two usual FV tests (see previous commit messages).

In SimulatorQueue.run(), the true stationary distribution was incorrectly computed by calling separately the compute_blocking_probability_birth_death_process() function on two different capacities: Ks-1 and Ks. This gave a stationary distribution on Ks-1 and Ks that could sum up to more than one in extreme cases (e.g. when Ks = 1, making Ks-1 and Ks the only two states of the system). The solution was to call stationary_distribution_product_form() on a system with capacity equal to Ks and keeping the probability values for states Ks-1 and Ks.

Note that this fix should ONLY affect the policy learning by FVRL (i.e. not the FV estimation of the blocking probability or expected cost), in particular when the true stationary probability is used instead of being estimated, which is used to provide a benchmark against which FVRL and Monte-Carlo learning can both be compared.

The tests in test_optimizers_continuoustime.py were updated with the new results (which do NOT affect the learning process but affect the reported true stationary probability and the error of the estimated stationary probability --see df_learning data frame created therein).