[FVRL] Set the minimum possible value for theta to THETA_MIN=0.1 instead of -0.9
Goal: Prevent admission control policy learners for queues and network systems from getting stuck at a negative theta value due to a persistent zero gradient, triggered by the fact that, when theta < 0, the deterministic blocking size K = 1, making the non-deterministic blocking state K-1 = 0, which is also equal to J-1. I.e. the non-deterministic blocking size K-1, which is the ONLY ONE contributing to the policy gradient, belongs to the absorption set and therefore it is NOT estimated by the Fleming-Viot process because FV can only estimate the probability of states OUTSIDE the absorption set A.
This problem is observed both in unidimensional control problems such as the single server queue and multidimensional control problem, such as a multi-class loss network.

This is a partial commit, so correct execution is not guaranteed.
Nevertheless, all unit tests should pass, except for the two usual FV tests (see previous commit messages).

DETAILS:
- agents/learners/policies.py: changed lower bound for theta from `THETA_MIN = -1 + 0.1` --> `THETA_MIN = 0.1`.

- environments/queues.py:
  - costBlockingExponential(): Updated documentation and comments so that they correspond to the exponential cost that is actually implemented (as opposed to the piecewise exponential cost, which was the original implementation that is constant before buffer_size_ref and then increases exponentially with the blocking size).

- plot_FVRL.py: Improved the plots so that labels are now written in Latex (e.g. $\theta_1$) and the minimum theta allowed for the policy learning process being plotted may be different from the previously hard-coded value of `-1 + 0.1 = -0.9` (as now the THETA_MIN value has been changed from -0.9 to 0.1).

- test/test_optimizers_continuoustime.py: updated the test that is affected by this change, namely the one where theta goes at one point below zero: test_Env_MetFVRL_TestSeveralCases().